+++ app/gui/src/dashboard/layouts/Settings/Settings.tsx
@@ -7,16 +7,14 @@ import { useStrictPortalContext } from '#/components/Portal'
 import { Text } from '#/components/Text'
 import { backendMutationOptions, backendQueryOptions } from '#/hooks/backendHooks'
 import { useEventCallback } from '#/hooks/eventCallbackHooks'
+import { useLocalStorageState } from '#/hooks/localStoreState'
 import { useSearchParamsState } from '#/hooks/searchParamsStateHooks'
 import { useToastAndLog } from '#/hooks/toastAndLogHooks'
 import SearchBar from '#/layouts/SearchBar'
-import { useFullUserSession } from '#/providers/AuthProvider'
-import { useLocalStorageState } from '#/providers/LocalStorageProvider'
-import { useSessionAPI } from '#/providers/SessionProvider'
 import { Path } from '#/services/ProjectManager'
 import { includesPredicate } from '#/utilities/array'
 import { regexEscape } from '#/utilities/string'
-import { useBackends, useText } from '$/providers/react'
+import { useBackends, useFullUserSession, useSession, useText } from '$/providers/react'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 import * as React from 'react'
 import {
@@ -42,7 +40,7 @@ export function Settings() {
     includesPredicate(Object.values(SettingsTabType)),
   )
   const { user, accessToken } = useFullUserSession()
-  const { changePassword } = useSessionAPI()
+  const { changePassword } = useSession()
   const { getText } = useText()
   const toastAndLog = useToastAndLog()
   const [query, setQuery] = React.useState('')