+++ app/gui/src/project-view/components/GraphEditor.vue
@@ -7,19 +7,13 @@ import {
   useWidgetRegistry,
 } from '$/components/WithCurrentProject.vue'
 import { useRightPanelData } from '$/providers/rightPanel'
-import {
-  codeEditorBindings,
-  documentationEditorBindings,
-  graphBindings,
-  undoBindings,
-} from '@/bindings'
+import { graphBindings, panelsBindings, undoBindings } from '@/bindings'
 import BottomPanel from '@/components/BottomPanel.vue'
 import CodeEditor from '@/components/CodeEditor.vue'
 import ComponentBrowser from '@/components/ComponentBrowser.vue'
 import type { Usage } from '@/components/ComponentBrowser/input'
 import { usePlacement } from '@/components/ComponentBrowser/placement'
 import ContextMenuTrigger from '@/components/ContextMenuTrigger.vue'
-import DocumentationEditor from '@/components/DocumentationEditor.vue'
 import GraphEdges from '@/components/GraphEditor/GraphEdges.vue'
 import GraphNodes from '@/components/GraphEditor/GraphNodes.vue'
 import { useGraphEditorClipboard } from '@/components/GraphEditor/clipboard'
@@ -34,7 +28,7 @@ import SceneScroller from '@/components/SceneScroller.vue'
 import TopBar from '@/components/TopBar.vue'
 import { builtinWidgets } from '@/components/widgets'
 import { useDoubleClick } from '@/composables/doubleClick'
-import { keyboardBusy, keyboardBusyExceptIn, unrefElement, useEvent } from '@/composables/events'
+import { keyboardBusy, unrefElement, useEvent } from '@/composables/events'
 import type { PlacementStrategy } from '@/composables/nodeCreation'
 import { ActionName, registerHandlers, toggledAction } from '@/providers/action'
 import { provideGraphEditorState } from '@/providers/graphEditorState'
@@ -61,19 +55,18 @@ import { Err, Ok, unwrapOr } from '@/util/data/result'
 import { Vec2 } from '@/util/data/vec2'
 import { isDef, VueInstance } from '@vueuse/core'
 import * as iter from 'enso-common/src/utilities/data/iter'
+import * as objects from 'enso-common/src/utilities/data/object'
 import { set } from 'lib0'
 import {
   computed,
   onMounted,
   onUnmounted,
   ref,
-  shallowRef,
   toRaw,
   toRef,
   useTemplateRef,
   watch,
   watchEffect,
-  type ComponentInstance,
 } from 'vue'
 
 const keyboard = injectKeyboard()
@@ -128,10 +121,9 @@ function selectionBounds() {
   return nodesBounds(nodeSelection.selected) ?? scrollBounds.value
 }
 
-function zoomToSelected(skipAnimation: boolean = false) {
+function zoomToSelected() {
   const bounds = selectionBounds()
-  if (bounds)
-    graphNavigator.panAndZoomTo(bounds, 0.1, Math.max(1, graphNavigator.targetScale), skipAnimation)
+  if (bounds) graphNavigator.panAndZoomTo(bounds, 0.1, Math.max(1, graphNavigator.targetScale))
 }
 
 function zoomToAll(skipAnimation: boolean = false) {
@@ -249,15 +241,15 @@ const actionHandlers = registerHandlers({
     action: () => nodeExecution.recomputeAll('Live'),
   },
   'graph.undo': {
+    enabled: graphStore.undoManager.canUndo,
     action: () => graphStore.undoManager.undo(),
-    disabled: () => !graphStore.undoManager.canUndo,
   },
   'graph.redo': {
+    enabled: graphStore.undoManager.canRedo,
     action: () => graphStore.undoManager.redo(),
-    disabled: () => !graphStore.undoManager.canRedo,
   },
   'graph.fitAll': {
-    action: () => zoomToSelected(),
+    action: zoomToSelected,
   },
   'graph.zoomIn': {
     action: () => graphNavigator.stepZoom(+1),
@@ -266,9 +258,9 @@ const actionHandlers = registerHandlers({
     action: () => graphNavigator.stepZoom(-1),
   },
   'graph.navigateUp': {
+    available: stackNavigator.hasBreadcrumbsBeyondRoot,
+    enabled: stackNavigator.allowNavigationLeft,
     action: () => stackNavigator.exitNode(),
-    disabled: () => !stackNavigator.allowNavigationLeft.value,
-    hidden: () => !stackNavigator.hasBreadcrumbsBeyondRoot.value,
   },
   ...selectionActionHandlers(
     () =>
@@ -291,10 +283,9 @@ useEvent(
   window,
   'keydown',
   (event) =>
+    panelsHandler(event) ||
     (!keyboardBusy() && undoBindingsHandler(event)) ||
     (!keyboardBusy() && graphBindingsHandler(event)) ||
-    (!keyboardBusyExceptIn(codeEditorArea.value) && codeEditorHandler(event)) ||
-    (!keyboardBusyExceptIn(documentationEditorArea.value) && documentationEditorHandler(event)) ||
     (!keyboardBusy() && graphNavigator.keyboardEvents.keydown(event)),
 )
 
@@ -320,12 +311,23 @@ const { handleClick } = useDoubleClick(
 
 // === Keyboard/Mouse bindings ===
 
-const undoBindingsHandler = undoBindings.handler({
-  undo: actionHandlers['graph.undo'].action,
-  redo: actionHandlers['graph.redo'].action,
-})
+const undoBindingsHandler = undoBindings.handler(
+  objects.mapEntries(undoBindings.bindings, (actionName) => actionHandlers[actionName].action),
+)
 
 const graphBindingsHandler = graphBindings.handler({
+  ...objects.unsafeFromEntries<Partial<Record<ActionName, () => void>>>(
+    (
+      [
+        'components.deleteSelected',
+        'components.copy',
+        'components.collapse',
+        'components.pickColorMulti',
+        'graph.fitAll',
+        'graph.navigateUp',
+      ] as const
+    ).map((actionName) => [actionName, actionHandlers[actionName].action]),
+  ),
   startProfiling() {
     projectStore.lsRpcConnection.profilingStart(true)
   },
@@ -336,10 +338,6 @@ const graphBindingsHandler = graphBindings.handler({
     if (graphNavigator.sceneMousePos != null && !componentBrowserOpened.value) {
       createWithComponentBrowser(fromSelection() ?? { placement: { type: 'mouse' } })
     }
-  },
-  deleteSelected: actionHandlers['components.deleteSelected'].action,
-  zoomToSelected() {
-    zoomToSelected()
   },
   selectAll() {
     nodeSelection.selectAll()
@@ -361,22 +359,14 @@ const graphBindingsHandler = graphBindings.handler({
       }
     })
   },
-  copyNode: actionHandlers['components.copy'].action,
   pasteNode() {
     createNodesFromClipboard()
   },
-  collapse: actionHandlers['components.collapse'].action,
   enterNode() {
     const selectedNode = set.first(nodeSelection.selected)
     if (selectedNode) {
       stackNavigator.enterNode(selectedNode)
     }
-  },
-  exitNode() {
-    stackNavigator.exitNode()
-  },
-  changeColorSelectedNodes() {
-    actionHandlers['components.pickColorMulti'].toggled.value = true
   },
   openDocumentation() {
     const result = tryGetSelectionDocUrl()
@@ -390,14 +380,10 @@ const graphBindingsHandler = graphBindings.handler({
 
 // === Code Editor ===
 
-const codeEditor = shallowRef<ComponentInstance<typeof CodeEditor>>()
-const codeEditorArea = computed(() => unrefElement(codeEditor))
 const showCodeEditor = ref(false)
-const codeEditorHandler = codeEditorBindings.handler({
-  toggle() {
-    showCodeEditor.value = !showCodeEditor.value
-  },
-})
+const panelsHandler = panelsBindings.handler(
+  objects.mapEntries(panelsBindings.bindings, (actionName) => actionHandlers[actionName].action),
+)
 
 // === Documentation Editor ===
 
@@ -425,13 +411,6 @@ watchEffect(() => {
 function toggleRightDockHelpPanel() {
   rightPanel.tab = 'help'
 }
-
-const docEditor = shallowRef<ComponentInstance<typeof DocumentationEditor>>()
-const documentationEditorArea = computed(() => unrefElement(docEditor))
-
-const documentationEditorHandler = documentationEditorBindings.handler({
-  toggle: () => rightPanel.toggleTab('documentation'),
-})
 
 // === Component Browser ===
 
@@ -703,12 +682,7 @@ const contextMenuActions: ActionName[] = [
           />
         </template>
         <TopBar
-          v-model:recordMode="projectStore.recordMode"
-          v-model:showCodeEditor="showCodeEditor"
           v-model:projectNameEdited="projectNameEdited"
-          v-model:showDocumentationEditor="
-            actionHandlers['graph.toggleDocumentationEditor'].toggled
-          "
           :zoomLevel="100.0 * graphNavigator.targetScale"
           :menuActions="contextMenuActions"
           @contextmenu.stop.prevent
@@ -717,7 +691,7 @@ const contextMenuActions: ActionName[] = [
         <GraphMouse />
       </ContextMenuTrigger>
       <BottomPanel v-model:show="showCodeEditor">
-        <CodeEditor ref="codeEditor" />
+        <CodeEditor />
       </BottomPanel>
     </div>
   </div>