+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/binary/LogicalOperations.java
@@ -4,10 +4,10 @@ import java.util.BitSet;
 import org.enso.table.data.column.builder.BoolBuilder;
 import org.enso.table.data.column.operation.BinaryOperation;
 import org.enso.table.data.column.operation.BinaryOperationBoolean;
-import org.enso.table.data.column.operation.map.MapOperationProblemAggregator;
 import org.enso.table.data.column.storage.BoolStorage;
 import org.enso.table.data.column.storage.ColumnBooleanStorage;
 import org.enso.table.data.column.storage.ColumnStorage;
+import org.enso.table.data.table.problems.MapOperationProblemAggregator;
 import org.enso.table.util.BitSets;
 
 /**