+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/comparators/GenericComparators.java
@@ -6,11 +6,11 @@ import org.enso.table.data.column.builder.BoolBuilder;
 import org.enso.table.data.column.builder.Builder;
 import org.enso.table.data.column.operation.BinaryOperation;
 import org.enso.table.data.column.operation.StorageIterators;
-import org.enso.table.data.column.operation.map.MapOperationProblemAggregator;
 import org.enso.table.data.column.storage.ColumnStorage;
 import org.enso.table.data.column.storage.type.AnyObjectType;
 import org.enso.table.data.column.storage.type.NullType;
 import org.enso.table.data.column.storage.type.StorageType;
+import org.enso.table.data.table.problems.MapOperationProblemAggregator;
 
 public class GenericComparators<T> implements BinaryOperation<Boolean> {
   protected final StorageType<T> valueType;